move path functions and make public
