+++ qcodes/utils/helpers.py
@@ -9,7 +9,6 @@ from contextlib import contextmanager
 from copy import deepcopy
 from functools import partial
 from inspect import signature
-from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -39,11 +38,11 @@ from ruamel.yaml import YAML
 
 from .attribute_helpers import DelegateAttributes, checked_getattr, strip_attrs
 from .json_utils import NumpyJSONEncoder
+from .path_helpers import QCODES_USER_PATH_ENV, get_qcodes_path, get_qcodes_user_path
 
 if TYPE_CHECKING:
     from PyQt5.QtWidgets import QMainWindow
 
-QCODES_USER_PATH_ENV = 'QCODES_USER_PATH'
 
 
 _tprint_times: Dict[str, float] = {}
@@ -538,29 +537,3 @@ def abstractmethod(funcobj: Callable[..., Any]) -> Callable[..., Any]:
     """
     funcobj.__qcodes_is_abstract_method__ = True  # type: ignore[attr-defined]
     return funcobj
-
-
-def get_qcodes_path(*subfolder: str) -> str:
-    """
-    Return full file path of the QCoDeS module. Additional arguments will be
-    appended as subfolder.
-
-    """
-    import qcodes
-    path = os.sep.join(qcodes.__file__.split(os.sep)[:-1])
-    return os.path.join(path, *subfolder) + os.sep
-
-
-def get_qcodes_user_path(*file_parts: str) -> str:
-    """
-    Get ``~/.qcodes`` path or if defined the path defined in the
-    ``QCODES_USER_PATH`` environment variable.
-
-    Returns:
-        path to the user qcodes directory
-
-    """
-    path = os.environ.get(QCODES_USER_PATH_ENV,
-                          os.path.join(Path.home(), '.qcodes'))
-    os.makedirs(os.path.dirname(path), exist_ok=True)
-    return os.path.join(path, *file_parts)
